examples/schelling: Restructure
- Split of SchellingAgent in agents.py
- Remove run_ascii.py (not supported anymore)
- Clean up files
