--- examples/basic/schelling/requirements.txt
@@ -1,3 +1,0 @@
-jupyter
-matplotlib
-mesa[viz]>=3.0.0b0
